AspNet vNext baseline
