+++ src/Orleans/Orleans.csproj
@@ -109,8 +109,6 @@
     <Compile Include="Providers\DependencyResolverProviderManager.cs" />
     <Compile Include="Providers\IDependencyResolverProvider.cs" />
     <Compile Include="Providers\IControllable.cs" />
-    <Compile Include="Providers\DependencyResolverProviderManager.cs" />
-    <Compile Include="Providers\IDependencyResolverProvider.cs" />
     <Compile Include="Providers\StorageProviderUtils.cs" />
     <Compile Include="RelationalStorage\AdoNetInvariants.cs" />
     <Compile Include="RelationalStorage\QueryConstantsBag.cs" />
