+++ src/Orleans/Providers/IDependencyResolverProvider.cs
@@ -23,14 +23,13 @@ TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR TH
 
 using Orleans.Runtime;
 using Orleans.Runtime.Configuration;
-using Orleans.Streams;
 
 namespace Orleans.Providers
 {
     /// <summary>
     /// Interface to be implemented by any app dependency resolver classes that want to be loaded and initialized during silo startup
     /// </summary>
-    internal interface IDependencyResolverProvider : IProvider
+    public interface IDependencyResolverProvider : IProvider
     {
         IDependencyResolver GetDependencyResolver(ClusterConfiguration config, NodeConfiguration nodeConfig, TraceLogger logger);
     }
