--- src/Orleans/project.json
@@ -1,67 +1,0 @@
-﻿{
-    "userSecretsId": "aspnet5-Microsoft.Orleans.Core-20150707052440",
-    "version": "2.0.0-*",
-    "description": "Core library of Microsoft Orleans used both on the client and server.",
-    "authors": [ "Microsoft Research" ],
-    "tags": [ "Orleans", "Cloud-Computing", "Actor-Model", "Actors", "Distributed-Systems", "C#", ".NET" ],
-    "projectUrl": "https://github.com/dotnet/Orleans",
-    "licenseUrl": "https://github.com/dotnet/Orleans#license",
-
-
-
-    "dependencies": {
-        "System.Reflection.Context": "4.0.0-beta-23019"
-    },
-    "exclude": "Properties\\orleans.codegen.cs",
-    "frameworks": {
-        "dnx451": {
-            "dependencies": {
-            },
-            "frameworkAssemblies": {
-                "System.Data": "4.0.0.0",
-                "System.Xml": "4.0.0.0",
-                "System.Management": "4.0.0.0",
-                "System.Transactions": "4.0.0.0"
-            }
-        },
-        "dnxcore50": {
-            "dependencies": {
-                "Microsoft.CSharp": "4.0.0-*",
-                "System.Collections": "4.0.10-*",
-                "System.Collections.Concurrent": "4.0.10-*",
-                "System.Console": "4.0.0-*",
-                "System.Data.SqlClient": "4.0.0-*",
-                "System.Diagnostics.Contracts": "4.0.0-*",
-                "System.Diagnostics.FileVersionInfo": "4.0.0-*",
-                "System.Diagnostics.Process": "4.0.0-*",
-                "System.Diagnostics.TraceSource": "4.0.0-*",
-                "System.Diagnostics.Tracing": "4.0.20-*",
-                "System.IO": "4.0.0-*",
-                "System.Linq": "4.0.0-*",
-                "System.Net.NameResolution": "4.0.0-*",
-                "System.Net.Primitives": "4.0.10-*",
-                "System.Net.Sockets": "4.0.10-*",
-                "System.ObjectModel": "4.0.10-*",
-                "System.Reflection": "4.0.10-*",
-                "System.Reflection.Extensions": "4.0.0-*",
-                "System.Reflection.Metadata": "1.0.22-*",
-                "System.Reflection.Primitives": "4.0.0-*",
-                "System.Reflection.TypeExtensions": "4.0.0-*",
-                "System.Runtime": "4.0.20-*",
-                "System.Runtime.Extensions": "4.0.10-*",
-                "System.Runtime.InteropServices": "4.0.20-*",
-                "System.Runtime.Serialization.Primitives": "4.0.10-*",
-                "System.Runtime.Serialization.Xml": "4.0.10-*",
-                "System.Security.Cryptography.Hashing": "4.0.0-*",
-                "System.Security.Cryptography.Hashing.Algorithms": "4.0.0-*",
-                "System.Threading": "4.0.10-*",
-                "System.Threading.Tasks": "4.0.10-*",
-                "System.Threading.ThreadPool": "4.0.10-*",
-                "System.Xml.XmlDocument": "4.0.0-beta-*",
-                "System.Xml.XmlSerializer": "4.0.10-beta-*"
-            }
-        }
-
-
-    }
-}
