+++ src/OrleansRuntime/DependencyInjection/DependencyResolverProviderBase.cs
@@ -1,0 +1,64 @@
+﻿/*
+Project Orleans Cloud Service SDK ver. 1.0
+ 
+Copyright (c) Microsoft Corporation
+ 
+All rights reserved.
+ 
+MIT License
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of this software and 
+associated documentation files (the ""Software""), to deal in the Software without restriction,
+including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
+and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so,
+subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS
+OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+
+using Orleans.Providers;
+using Orleans.Runtime.Configuration;
+using Orleans.Runtime.Management;
+using Orleans.Runtime.MembershipService;
+using Orleans.Runtime.ReminderService;
+using Orleans.Streams;
+
+namespace Orleans.Runtime
+{
+    public abstract class DependencyResolverProviderBase : IDependencyResolverProvider
+    {
+        public string Name
+        {
+            get { return "Orleans DependencyResolver Provider"; }
+        }
+
+        public System.Threading.Tasks.Task Init(string name, IProviderRuntime providerRuntime, IProviderConfiguration config)
+        {
+            //
+            // This method will not be called by the DependenceResolverProviderManager, when loading the provider.
+            //
+
+            throw new System.NotSupportedException();
+        }
+
+        public abstract IDependencyResolver ConfigureResolver(params System.Type[] systemTypesToRegister);
+
+        public IDependencyResolver GetDependencyResolver(ClusterConfiguration config, NodeConfiguration nodeConfig, TraceLogger logger)
+        {
+            var dependencyResolver = ConfigureResolver(
+                typeof(ManagementGrain),
+                typeof(GrainBasedMembershipTable),
+                typeof(GrainBasedReminderTable),
+                typeof(PubSubRendezvousGrain),
+                //typeof(MemoryStorageGrain), // Uncomment when moved to OrleansRuntime
+                typeof(GrainBasedPubSubRuntime));
+
+            return dependencyResolver;
+        }
+    }
+}
